fix(phase-17): resolve all 26 clippy -D warnings errors
- build.rs: use strip_prefix() instead of manual index slicing
- Cargo.toml: declare esp_idf_log_timestamp_source_system as known cfg
- log_relay.rs: remove redundant #[allow] (now handled via Cargo.toml)
- mqtt.rs: m @ _ → m; add #[allow(too_many_arguments)]
- provisioning.rs: remove redundant mut bindings; resp[2] |= 0x80
- wifi.rs: #[allow(dead_code)] on wifi_connect
- main.rs: fix doc comment formatting (doc_lazy_continuation)
- device_id.rs: remove redundant ESP_OK as i32 cast
- gnss.rs: #[allow(type_complexity)]; remove redundant usize cast
- ntrip_client.rs: div_ceil, Error::other, if not match for bool
- uart_bridge.rs: collapse if into match guards

+++ src/device_id.rs
@@ -23,7 +23,7 @@ pub fn get() -> String {
     let ret = unsafe { esp_efuse_mac_get_default(mac.as_mut_ptr()) };
     assert_eq!(
         ret,
-        ESP_OK as i32,
+        ESP_OK,
         "esp_efuse_mac_get_default failed: err={}",
         ret
     );
